+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Sync to upstream Luau 0.595
 - Prioritise `game:GetService()` as the first autocompletion entry when typing `game:`
 
+### Fixed
+
+- Do not add line separator in hover when there is no text documentation
+
 ## [1.24.1] - 2023-09-09
 
 ### Changed